Make data actually work
The main moves into a fatpacked file too, this is because 'tell DATA'
can't work during BEGIN blocks, as perl hasn't seen the __DATA__ to
create the filehandle.

I'm not too keen on this, it probably needs something like extracting
the pod from main so that perldoc fatpacked-file works correctly, it
breaks current cpanm as that relies on 'caller' returning false in main,
etc.