Stop chdir when finding packlist. Absolutify %INC path in case there's a relative path in @INC.
When there's a relative path in @INC (e.g. with -Mlib=./local/lib/perl5),
FatPacker died with `Can't open local/lib/perl5/.../.packlist` even if the
file exists, because of File::Find's chdir, and $File::Find::name is yet
relative.

This patch adds 'no_chdir' option to File::Find to stop chdir'ing to
the directories, and then expand the %INC entries to fullpath with
Cwd::abs_path, so that it will hopefully match with what you have in
.packlist content.

Reproduce:

  cpanm -l local Plack
  perl -Mlib=local/lib/perl5 -S fatpack packlists-for Plack.pm